+++ omicverse/single/__init__.py
@@ -11,3 +11,5 @@ from ._cpdb import cpdb_network_cal,cpdb_plot_network,cpdb_plot_interaction,cpdb
 from ._scgsea import geneset_aucell,pathway_aucell,pathway_aucell_enrichment,pathway_enrichment,pathway_enrichment_plot
 from ._via import pyVIA,scRNA_hematopoiesis
 from ._simba import pySIMBA
+from ._tosica import pyTOSICA
+from ._atac import atac_concat_get_index,atac_concat_inner,atac_concat_outer