+++ omicverse/tosica/__init__.py
@@ -1,0 +1,10 @@
+r"""
+TOSICA (Transformer for One-Stop Interpretable Cell-type Annotation)
+"""
+
+
+from . import train, pre, TOSICA_model
+from .train import fit_model
+from .pre import predict
+import os
+